requirements.txt
Flask
PyPDF2
reportlab
python-docx
boto3